requirements.txt
streamlit
PyMuPDF
Pillow
openai
python-dotenv
